introduced AdminClientProvider to make it easier to use
fixes: #48623
