+++ js/apps/admin-ui/src/App.tsx
@@ -1,20 +1,15 @@
-import KeycloakAdminClient from "@keycloak/keycloak-admin-client";
-import {
-  mainPageContentId,
-  useEnvironment,
-} from "@keycloak/keycloak-ui-shared";
-import { Flex, FlexItem, Page } from "@patternfly/react-core";
-import { PropsWithChildren, Suspense, useEffect, useState } from "react";
-import { Outlet } from "react-router-dom";
-
 import {
   ErrorBoundaryFallback,
   ErrorBoundaryProvider,
   KeycloakSpinner,
+  mainPageContentId,
 } from "@keycloak/keycloak-ui-shared";
+import { Flex, FlexItem, Page } from "@patternfly/react-core";
+import { PropsWithChildren, Suspense, useEffect } from "react";
+import { Outlet } from "react-router-dom";
+import { AdminClientProvider } from "./admin-client";
 import { Header } from "./PageHeader";
 import { PageNav } from "./PageNav";
-import { AdminClientContext, initAdminClient } from "./admin-client";
 import { PageBreadCrumbs } from "./components/bread-crumb/PageBreadCrumbs";
 import { ErrorRenderer } from "./components/error/ErrorRenderer";
 import { RecentRealmsProvider } from "./context/RecentRealms";
@@ -46,25 +41,14 @@ export const AppContexts = ({ children }: PropsWithChildren) => (
 );
 
 export const App = () => {
-  const { keycloak, environment } = useEnvironment<Environment>();
-  const [adminClient, setAdminClient] = useState<KeycloakAdminClient>();
-
+  const hrefEndsWithHashSlash = location.href.endsWith("#/");
   useEffect(() => {
-    const fragment = "#/";
-    if (window.location.href.endsWith(fragment)) {
-      const newPath = window.location.pathname.replace(fragment, "");
-      window.history.replaceState(null, "", newPath);
-    }
-    const init = async () => {
-      const client = await initAdminClient(keycloak, environment);
-      setAdminClient(client);
-    };
-    init().catch(console.error);
-  }, [environment, keycloak]);
+    if (!hrefEndsWithHashSlash) return;
+    history.replaceState(null, "", location.pathname);
+  }, [hrefEndsWithHashSlash, location.pathname]);
 
-  if (!adminClient) return <KeycloakSpinner />;
   return (
-    <AdminClientContext.Provider value={{ keycloak, adminClient }}>
+    <AdminClientProvider>
       <AppContexts>
         <Flex
           direction={{ default: "column" }}
@@ -94,6 +78,6 @@ export const App = () => {
           </FlexItem>
         </Flex>
       </AppContexts>
-    </AdminClientContext.Provider>
+    </AdminClientProvider>
   );
 };
